// src/enums.ts
export enum NodeEnvEnum {
  PRODUCTION = 'PRODUCTION',
  DEVELOPMENT = 'DEVELOPMENT',
  TESTING = 'TESTING',
}

export enum EnvEnum {
  NODE_ENV = 'NODE_ENV',
  PORT = 'PORT',
  TELEGRAM_BOT_TOKEN = 'TELEGRAM_BOT_TOKEN',
  TELEGRAM_BOT_DOMAIN = 'TELEGRAM_BOT_DOMAIN',
  TELEGRAM_BOT_HOOK_PATH = 'TELEGRAM_BOT_HOOK_PATH',
}